added new api module for creating custom rules and moved error classes to new module.
