+++ api/src/main/java/org/openapitools/openapistylevalidator/api/IRuleManager.java
@@ -1,0 +1,8 @@
+package org.openapitools.openapistylevalidator.api;
+
+import java.util.List;
+
+public interface IRuleManager {
+
+    List<IRule> loadRules();
+}
