+++ api/src/main/java/org/openapitools/openapistylevalidator/error/GenericStyleError.java
@@ -1,4 +1,4 @@
-package org.openapitools.openapistylevalidator.styleerror;
+package org.openapitools.openapistylevalidator.error;
 
 import java.util.Objects;
