+++ api/src/main/java/org/openapitools/openapistylevalidator/error/OperationStyleError.java
@@ -1,20 +1,15 @@
-package org.openapitools.openapistylevalidator.styleerror;
+package org.openapitools.openapistylevalidator.error;
 
 import java.util.Objects;
 import org.eclipse.microprofile.openapi.models.PathItem;
 
-public final class OperationNamingStyleError extends StyleError {
+public final class OperationStyleError extends StyleError {
 
     private final String path;
     private final PathItem.HttpMethod method;
 
-    public OperationNamingStyleError(
-            StyleCheckSection styleCheckSection,
-            String fieldNames,
-            String description,
-            String path,
-            PathItem.HttpMethod method) {
-        super(styleCheckSection, fieldNames, description);
+    public OperationStyleError(String fieldNames, String description, String path, PathItem.HttpMethod method) {
+        super(StyleCheckSection.Operations, fieldNames, description);
         this.path = path;
         this.method = method;
     }
@@ -23,7 +18,7 @@ public final class OperationNamingStyleError extends StyleError {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        OperationNamingStyleError that = (OperationNamingStyleError) o;
+        OperationStyleError that = (OperationStyleError) o;
         return styleCheckSection == that.styleCheckSection
                 && Objects.equals(fieldNames, that.fieldNames)
                 && Objects.equals(description, that.description)
@@ -38,8 +33,6 @@ public final class OperationNamingStyleError extends StyleError {
 
     @Override
     public String toString() {
-        return String.format(
-                "*ERROR* in path %s%s '%s' -> %s",
-                method == null ? "" : String.format("%s ", method.name()), path, fieldNames, description);
+        return String.format("*ERROR* in Operation %s %s '%s' -> %s", method.name(), path, fieldNames, description);
     }
 }
