+++ api/src/main/java/org/openapitools/openapistylevalidator/error/StyleCheckSection.java
@@ -1,0 +1,9 @@
+package org.openapitools.openapistylevalidator.error;
+
+public enum StyleCheckSection {
+    APIInfo,
+    Operations,
+    Models,
+    Naming,
+    OpenAPI,
+}
