+++ cli/build.gradle
@@ -6,6 +6,7 @@ apply plugin: 'application'
 
 dependencies {
     implementation project(':openapi-style-validator-lib')
+    implementation project(':openapi-style-validator-api')
     implementation 'commons-cli:commons-cli:1.5.0'
     implementation 'org.slf4j:slf4j-jdk14:2.0.0'
     implementation 'io.swagger.parser.v3:swagger-parser:2.0.33'
