+++ cli/src/main/java/org/openapitools/openapistylevalidator/cli/ValidationInitiator.java
@@ -9,7 +9,7 @@ import org.eclipse.microprofile.openapi.models.OpenAPI;
 import org.openapitools.empoa.swagger.core.internal.SwAdapter;
 import org.openapitools.openapistylevalidator.OpenApiSpecStyleValidator;
 import org.openapitools.openapistylevalidator.ValidatorParameters;
-import org.openapitools.openapistylevalidator.styleerror.StyleError;
+import org.openapitools.openapistylevalidator.error.StyleError;
 
 public class ValidationInitiator {
     public List<StyleError> validate(OptionManager optionManager, CommandLine commandLine) {
