+++ gradle-plugin/build.gradle
@@ -4,6 +4,7 @@ plugins {
 
 dependencies {
     implementation project(':openapi-style-validator-lib')
+    implementation project(':openapi-style-validator-api')
     implementation 'io.swagger.parser.v3:swagger-parser:2.0.33'
     implementation 'org.openapitools.empoa:empoa-swagger-core:2.0.0'
     testImplementation 'junit:junit:4.13.2'
