+++ lib/src/test/java/org/openapitools/openapistylevalidator/OpenApiSpecStyleValidatorTest.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.openapitools.openapistylevalidator.ValidatorParameters.NamingConvention;
-import org.openapitools.openapistylevalidator.styleerror.StyleError;
+import org.openapitools.openapistylevalidator.error.StyleError;
 
 class OpenApiSpecStyleValidatorTest {
