+++ lib/src/test/java/org/openapitools/openapistylevalidator/ValidatorParametersTest.java
@@ -1,6 +1,8 @@
 package org.openapitools.openapistylevalidator;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
