+++ maven-plugin/build.gradle
@@ -11,7 +11,7 @@ dependencies {
         exclude group: 'javax.inject', module: 'javax.inject'
         exclude group: 'org.codehaus.plexus', module: 'plexus-component-annotations'
     }
-
+    implementation project(':openapi-style-validator-api')
     implementation project(':openapi-style-validator-lib')
     implementation 'io.swagger.parser.v3:swagger-parser:2.0.33'
     implementation 'org.openapitools.empoa:empoa-swagger-core:2.0.0'
