+++ settings.gradle
@@ -4,7 +4,10 @@ include(':openapi-style-validator-lib')
 include(':openapi-style-validator-cli')
 include(':openapi-style-validator-gradle-plugin')
 include(':openapi-style-validator-maven-plugin')
+include(':openapi-style-validator-api')
 project(":openapi-style-validator-lib").projectDir = file('lib')
 project(":openapi-style-validator-cli").projectDir = file('cli')
 project(":openapi-style-validator-gradle-plugin").projectDir = file('gradle-plugin')
 project(":openapi-style-validator-maven-plugin").projectDir = file('maven-plugin')
+project(":openapi-style-validator-api").projectDir = file('api')
+
